backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "saleemasekrea000",email = "s.asekrea@innopolis.university"}
]
packages = [{include = "app"}]

readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "bcrypt (==3.2.2)",
    "redis (>=5.2.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "celery (>=5.5.2,<6.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "flower (>=2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"
pytest = "^7.0"
pytest-asyncio = "^0.20"
pytest-mock = "^3.10"
hypothesis = "^6.80"
coverage = "^6.5"
python-dotenv = "^1.0"